Finishing Stock products association

+++ stock-service/src/main/java/tn/esprit/Entity/Offer.java
@@ -25,7 +25,7 @@ public class Offer implements Serializable {
     @NonNull
     String quantity;
     @NonNull
-    float prix;
+    Double prix;
 
     @ManyToOne
     Product product;
